Use int32_t for cluster, area, node->num_surfaces, etc.
The only remaining int16_t's in BSP are mins, maxs and texinfo references.

+++ src/client/renderer/r_types.h
@@ -611,8 +611,8 @@ typedef struct {
 	struct r_model_s *model;
 
 	// leaf specific
-	int16_t cluster;
-	int16_t area;
+	int32_t cluster;
+	int32_t area;
 
 	r_bsp_surface_t **first_leaf_surface;
 	int32_t num_leaf_surfaces;
